requirements.txt
# 本番用依存関係
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.103.1
uvicorn[standard]==0.23.2

# 開発用依存関係
pytest==7.4.2
pytest-asyncio==0.23.5
mypy==1.5.1
black==23.9.1
isort==5.12.0